Apply nightly Clippy suggestions

+++ bxt-strafe/src/steps.rs
@@ -527,7 +527,7 @@ impl<S: Step> Step for Jump<S> {
     ) -> (State, Input) {
         if parameters.has_stamina {
             state.player.stamina_time =
-                (state.player.stamina_time - (parameters.frame_time * 1000.) as f32).max(0.);
+                (state.player.stamina_time - parameters.frame_time * 1000.).max(0.);
         }
 
         if input.jump && !state.prev_frame_input.jump && state.place == Place::Ground {
